[WFCORE-5442] Math operands should be cast before assignment
Fixes https://issues.redhat.com/browse/WFCORE-5442

+++ patching/src/main/java/org/jboss/as/patching/runner/PatchModuleInvalidationUtils.java
@@ -392,7 +392,7 @@ class PatchModuleInvalidationUtils {
                 break;
             }
             lastChannelPos = channelPos;
-            channelPos -= Math.min(channelPos - bufferPos, CHUNK_SIZE - bufferPos);
+            channelPos -= Math.min(channelPos - bufferPos, (long) CHUNK_SIZE - bufferPos);
         }
 
         return -1;
